AI_Agent_Llamaindex/requirements.txt
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-google-genai>=0.1.11
llama-index-vector-stores-chroma>=0.1.0
python-dotenv>=0.19.0
tiktoken>=0.5.0
scikit-learn>=1.0.0
numpy>=1.20.0
pypdf>=3.0.0
pdfminer.six>=20221105
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
cohere==4.50